+++ lib/rapa/client.rb
@@ -22,16 +22,13 @@ module Rapa
       end
     end
 
+    # @param domain [String]
     # @param options [Hash]
     # @return [Rapa::Responses::ListItemsResponse]
-    def list_items(options)
-      query = ::Rapa::Queries::ListItemsQuery.new(
-        {
-          access_key_id: access_key_id,
-          associate_tag: associate_tag,
-        }.merge(options),
-      )
-      url = ::Rapa::Url.new(domain: options[:domain])
+    def list_items(domain:, **options)
+      query_options = common_query_options.merge(options)
+      query = ::Rapa::Queries::ListItemsQuery.new(query_options)
+      url = ::Rapa::Url.new(domain: domain)
       uri = ::URI.parse(url.to_s)
       query_string = query.to_s
       signature = ::Rapa::Signer.new(
@@ -43,6 +40,26 @@ module Rapa
       ).sign
       faraday_response = connection.get(uri, query.to_hash.merge(Signature: signature))
       ::Rapa::Responses::ListItemsResponse.new(faraday_response)
+    end
+
+    # @param domain [String]
+    # @param options [Hash]
+    # @return [Rapa::Responses::SearchItemsResponse]
+    def search_items(domain:, **options)
+      query_options = common_query_options.merge(options)
+      query = ::Rapa::Queries::SearchItemsQuery.new(query_options)
+      url = ::Rapa::Url.new(domain: domain)
+      uri = ::URI.parse(url.to_s)
+      query_string = query.to_s
+      signature = ::Rapa::Signer.new(
+        host: uri.host,
+        http_method: "GET",
+        key: secret_access_key,
+        path: uri.path,
+        query_string: query_string,
+      ).sign
+      faraday_response = connection.get(uri, query.to_hash.merge(Signature: signature))
+      ::Rapa::Responses::SearchItemsResponse.new(faraday_response)
     end
 
     private
@@ -57,6 +74,15 @@ module Rapa
     # @return [String]
     def associate_tag
       @associate_tag
+    end
+
+    # @private
+    # @return [Hash]
+    def common_query_options
+      {
+        access_key_id: access_key_id,
+        associate_tag: associate_tag,
+      }
     end
 
     # @private