+++ lib/rapa/responses/search_items_response.rb
@@ -1,0 +1,12 @@
+module Rapa
+  module Responses
+    class SearchItemsResponse < BaseResponse
+      private
+
+      # @note Override
+      def resource_class
+        ::Rapa::Resources::ItemResource
+      end
+    end
+  end
+end